+++ src/FFTPack/fftpack5.1/src/mcfti1.f
@@ -29,14 +29,14 @@ C     *                                                               *
 C     * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
 C
       SUBROUTINE MCFTI1 (N,WA,FNF,FAC)
-      REAL            WA(*),FAC(*) 
+      REAL               WA(*),FAC(*)
 C
       CALL FACTOR (N,NF,FAC)
       FNF = NF
       IW = 1
       L1 = 1
       DO 110 K1=1,NF
-         IP = FAC(K1)
+         IP = INT(FAC(K1))
          L2 = L1*IP
          IDO = N/L2
          CALL TABLES (IDO,IP,WA(IW))
